requirements.txt
asgiref==3.8.1
boto3==1.38.46
botocore==1.38.46
certifi==2025.6.15
charset-normalizer==3.4.2
Django==4.2.23
django-anymail==13.0
django-taggit==6.1.0
idna==3.10
jmespath==1.0.1
pillow==11.2.1
psycopg2-binary==2.9.10
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
requests==2.32.4
s3transfer==0.13.0
six==1.17.0
sqlparse==0.5.3
typing_extensions==4.14.0
tzdata==2025.2
urllib3==1.26.20
